Modified : modifies component name
